MLDBFB-615: split libcloud into libaws and libvfs_handlers
