+++ mldb.mk
@@ -1,3 +1,3 @@
-MLDB_SUBDIRS := ext arch base types watch logging io http vfs jml soa ml utils rest sql core builtin plugins server sdk tensorflow testing
+MLDB_SUBDIRS := ext arch base types watch logging io http vfs jml soa ml vfs_handlers utils rest sql core builtin plugins server sdk tensorflow testing
 
 $(eval $(call include_sub_makes,$(MLDB_SUBDIRS)))
