+++ server/server.mk
@@ -28,7 +28,7 @@ LIBMLDB_SOURCES:= \
 	bucket.cc \
 
 LIBMLDB_LINK:= \
-	service_peer mldb_builtin_plugins sql_expression runner credentials git2 hoedown mldb_builtin command_expression cloud mldb_core
+	service_peer mldb_builtin_plugins sql_expression runner credentials git2 hoedown mldb_builtin command_expression vfs_handlers mldb_core
 
 
 $(eval $(call library,mldb,$(LIBMLDB_SOURCES),$(LIBMLDB_LINK)))
