+++ vfs_handlers/hdfs.cc
