+++ vfs_handlers/s3_transfer_cmd.cc
