+++ vfs_handlers/s3cat.cc
