+++ vfs_handlers/s3cp.cc
