+++ vfs_handlers/s3tee.cc
