+++ vfs_handlers/testing/testing.mk
@@ -1,0 +1,2 @@
+#$(eval $(call test,hdfs_test,vfs_handlers,boost manual))
+$(eval $(call test,fs_utils_test,vfs_handlers test_utils,boost manual))
