+++ vfs_handlers/vfs_handlers.mk
@@ -1,0 +1,25 @@
+# VFS Handlers
+
+LIBVFS_HANDLERS_SOURCES := \
+	sftp.cc \
+	s3_handlers.cc \
+	archive.cc \
+	docker.cc
+
+#	hdfs.cc
+
+LIBVFS_HANDLERS_LINK := aws hash archive tinyxml2 ssh2  #hdfs3
+
+
+$(eval $(call library,vfs_handlers,$(LIBVFS_HANDLERS_SOURCES),$(LIBVFS_HANDLERS_LINK)))
+
+# gcc 4.7
+$(eval $(call set_compile_option,aws.cc,-fpermissive))
+
+$(eval $(call program,s3_transfer_cmd,vfs_handlers boost_program_options boost_filesystem utils))
+$(eval $(call program,s3tee,vfs_handlers boost_program_options utils))
+$(eval $(call program,s3cp,vfs_handlers boost_program_options utils))
+$(eval $(call program,s3_multipart_cmd,vfs_handlers boost_program_options utils))
+$(eval $(call program,s3cat,vfs_handlers boost_program_options utils))
+
+$(eval $(call include_sub_make,testing))
